+++ src/test/helpers.ts
@@ -8,8 +8,8 @@ import { TextDocument } from "vscode-languageserver-textdocument";
 import { URI } from "vscode-uri";
 import { parseDocument } from "../server/services/parser";
 import type StorageService from "../server/services/storage";
+import type { ISettings } from "../server/settings";
 import type { INode } from "../server/types/nodes";
-import type { ISettings } from "../server/types/settings";
 
 const ls = getSCSSLanguageService();
